Organized and docemented portions of code
