+++ src/App.vue
@@ -1,3 +1,14 @@
+
+<!-- TODO 
+ 
+- Add padding for the indoor scrollbar.
+- Investigate initial camera zoom on individual buildings.
+- Realign all the floors of Deerfield Hall
+- Find out what each CSS stle does
+- Move css styles to another file
+
+END OF TODO -->
+
 <template>
     <div id="app">
       
@@ -7,6 +18,7 @@
         <!--<img src="./assets/UofT.png">-->
 
         <!-- <h1>{{ msg }}</h1> -->
+        
         <div id="widget-container" class="wrld-widget-container"></div>
 
         <div id="map" style="height: 96vh;"></div>
@@ -14,6 +26,7 @@
         <link href="https://cdnjs.cloudflare.com/ajax/libs/leaflet/1.0.1/leaflet.css" rel="stylesheet" />
 
         <link href="https://cdn-webgl.wrld3d.com/wrldjs/addons/resources/latest/css/wrld.css" rel="stylesheet"/>
+        
         <!--<script src="https://code.jquery.com/jquery-1.12.4.min.js"></script>
         <script src="https://code.jquery.com/ui/1.12.0/jquery-ui.min.js"></script>
         <script src="https://cdn-webgl.wrld3d.com/wrldjs/addons/indocontrol/latest/indoor_control.js"></script>-->
@@ -30,31 +43,27 @@
 
 <script>
 
+/* All imports and requirements go here */
+
 import { initialCameraZoom }  from './initial_camera_zoom';
 import { onLoadScrollbar }  from './indoor_control';
-import { pad }  from './indoor_control'; // # TODO in progress, pads the document
+import { createMap }  from './create_map';
+// import { pad }  from './indoor_control'; // # TODO in progress, adds padding to the indoor scrollbar
 
-const Wrld = require("wrld.js")
+const Wrld = require("wrld.js");
+
+/* All Javascript functions go in mounted */
 
 export default {
     name: 'app',
     mounted() {
-        var map = L.Wrld.map("map", "5378c39112e718bdeb6f19df0168d1cf", {
-            center: [43.549, -79.6636],
-            zoom: 15,
-            trafficEnabled: false,
-            frameRateThrottleWhenIdleEnabled: true,
-            throttledTargetFrameIntervalMilliseconds: 500,
-            idleSecondsBeforeFrameRateThrottle: 15.0,
-            indoorsEnabled: true,
-            coverageTreeManifest: "https://webgl-cdn1.wrld3d.com/chunk/indoor_maps/api_requests/EIM-c3eb2f77-20e3-4b6b-bb11-784ced915fa0_2020_02_09_05_33_40/webgl_manifest.bin.gz"
-        });
+        var map = createMap();
 
-        // initialCameraZoom(map)
+        // initialCameraZoom(map);
 
-        onLoadScrollbar(map)
+        onLoadScrollbar(map);
 
-        // pad()
+        // pad();
     },
   
     data () {
@@ -64,6 +73,9 @@ export default {
     },
 }
 </script>
+
+<!-- All CSS styles
+-->
 
 <style>
 body {
